Remove typedefs from structs, and remove the `const` nonsense
Typedefs from structs may seem cool but there is no need to introduce a
global name. If the user wants to typedef, they can.

The "`const` everything" bandwagon doesn't make much of a sense in a
language with highly mutable state like C, I can simply take a pointer
to a `const` and modify it anyway. Also, adding `const` everywhere
decreases significantly the readability of the code and helps in
absolutely nothing in terms of performance.

Signed-off-by: Luiz Mugnaini <luizmugnaini@gmail.com>